header component and profile page updated
Header UI now changes based on if user is logged in or not. Profile now fetches data based on the current user that is logged in, will update the profileCl when database is updated. Users can now logout, and to navigate to login, no active session must be present. Similarly to navigate to profile page, an active session must be present.